fixed a few things

+++ src/Input/InputAction.h
@@ -8,7 +8,7 @@ using InputKey = uint8_t;
 using InputState = uint8_t;
 using MouseButton = uint8_t;
 
-using InputAction = std::function<void(uint32_t dt, InputState)>;
+using InputAction = std::function<void(uint64_t dt, InputState)>;
 
 struct ButtonAction
 {
